+++ src/platform.ts
@@ -12,41 +12,58 @@ import { PLATFORM_NAME, PLUGIN_NAME } from './settings';
 import { SpotifyApiWrapper } from './spotify-api-wrapper';
 import { SpotifySpeakerAccessory } from './spotify-speaker-accessory';
 import type { HomebridgeSpotifySpeakerDevice } from './types';
+import { PluginLogger } from './plugin-logger';
 
 const DEVICE_CLASS_CONFIG_MAP = {
   speaker: SpotifySpeakerAccessory,
 };
 
-const DAY_INTERVAL = 60 * 60 * 24 * 1000;
+const DAY_INTERVAL_MS = 60 * 60 * 24 * 1000;
+const MINUTE_INTERVAL_MS = 60 * 1000;
+const DEFAULT_POLL_INTERVAL_S = 20;
 
 export class HomebridgeSpotifySpeakerPlatform implements DynamicPlatformPlugin {
   public readonly Service: typeof Service = this.api.hap.Service;
   public readonly Characteristic: typeof Characteristic = this.api.hap.Characteristic;
+  public readonly logger: PluginLogger;
   public readonly spotifyApiWrapper: SpotifyApiWrapper;
+  public readonly pollIntervalSec: number;
   public readonly accessories: {
     [uuid: string]: PlatformAccessory;
   } = {};
 
-  constructor(public readonly log: Logger, public readonly config: PlatformConfig, public readonly api: API) {
-    this.log.debug('Finished initializing platform:', this.config.name);
+  constructor(readonly log: Logger, public readonly config: PlatformConfig, public readonly api: API) {
+    this.logger = new PluginLogger(log, config);
+    this.logger.debug('Finished initializing platform:', this.config.name);
 
     if (!config.spotifyClientId || !config.spotifyClientSecret || !config.spotifyAuthCode) {
-      this.log.error('Missing configuration for this plugin to work, see the documentation for initial setup');
+      this.logger.error('Missing configuration for this plugin to work, see the documentation for initial setup');
       return;
     }
+    this.pollIntervalSec = config.spotifyPollInterval || DEFAULT_POLL_INTERVAL_S;
 
-    this.spotifyApiWrapper = new SpotifyApiWrapper(log, config, api);
+    this.spotifyApiWrapper = new SpotifyApiWrapper(this.logger, config, api);
 
     this.api.on('didFinishLaunching', async () => {
-      log.debug('Executed didFinishLaunching callback');
+      this.logger.debug('Executed didFinishLaunching callback');
 
       const isAuthenticated = await this.spotifyApiWrapper.authenticate();
       if (!isAuthenticated) {
         return;
       }
 
-      this.logAvailableSpotifyDevices();
+      await this.logAvailableSpotifyDevices();
+      await this.setSpotifyPlaybackState();
       this.discoverDevices();
+      if (Object.keys(this.accessories).length) {
+        setInterval(() => {
+          this.setSpotifyDevices();
+        }, MINUTE_INTERVAL_MS);
+
+        setInterval(() => {
+          this.setSpotifyPlaybackState();
+        }, this.pollIntervalSec * 1000);
+      }
     });
 
     // Make sure we have the latest tokens saved
@@ -54,17 +71,19 @@ export class HomebridgeSpotifySpeakerPlatform implements DynamicPlatformPlugin {
       this.spotifyApiWrapper.persistTokens();
     });
 
-    setInterval(async () => await this.spotifyApiWrapper.refreshTokens(), DAY_INTERVAL);
+    setInterval(async () => await this.spotifyApiWrapper.refreshTokens(), DAY_INTERVAL_MS);
   }
 
   configureAccessory(accessory: PlatformAccessory) {
-    this.log.info('Loading accessory from cache:', accessory.displayName);
+    this.logger.info('Loading accessory from cache:', accessory.displayName);
     this.accessories[accessory.UUID] = accessory;
   }
 
   discoverDevices() {
     if (!this.config.devices) {
-      this.log.error('The "devices" section is missing in your plugin configuration, please add at least one device.');
+      this.logger.error(
+        'The "devices" section is missing in your plugin configuration, please add at least one device.',
+      );
       return;
     }
 
@@ -78,7 +97,7 @@ export class HomebridgeSpotifySpeakerPlatform implements DynamicPlatformPlugin {
         continue;
       }
       if (!this.deviceConfigurationIsValid(device)) {
-        this.log.error(
+        this.logger.error(
           `${
             device.deviceName ?? 'unknown device'
           } is not configured correctly. See the documentation for initial setup`,
@@ -96,13 +115,13 @@ export class HomebridgeSpotifySpeakerPlatform implements DynamicPlatformPlugin {
         existingAccessory ?? new this.api.platformAccessory(device.deviceName, uuid, deviceClass.CATEGORY);
       accessory.context.device = device;
       accessory.context.playlistId = playlistId;
-      new deviceClass(this, accessory, device, this.log);
+      new deviceClass(this, accessory, device, this.logger);
       activeAccessoryIds.push(uuid);
 
       if (existingAccessory) {
-        this.log.info('Restoring existing accessory from cache:', accessory.displayName);
+        this.logger.info('Restoring existing accessory from cache:', accessory.displayName);
       } else {
-        this.log.info('Adding new accessory:', device.deviceName);
+        this.logger.info('Adding new accessory:', device.deviceName);
         platformAccessories.push(accessory);
       }
     }
@@ -116,11 +135,29 @@ export class HomebridgeSpotifySpeakerPlatform implements DynamicPlatformPlugin {
       .map((id) => this.accessories[id]);
 
     staleAccessories.forEach((staleAccessory) => {
-      this.log.info(`Removing stale cached accessory ${staleAccessory.UUID} ${staleAccessory.displayName}`);
+      this.logger.info(`Removing stale cached accessory ${staleAccessory.UUID} ${staleAccessory.displayName}`);
     });
 
     if (staleAccessories.length) {
       this.api.unregisterPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, staleAccessories);
+    }
+  }
+
+  async setSpotifyDevices(): Promise<void> {
+    try {
+      const devices = await this.spotifyApiWrapper.getMyDevices();
+      SpotifySpeakerAccessory.DEVICES = devices;
+    } catch {
+      this.logger.error('Error setting spotify devices');
+    }
+  }
+
+  async setSpotifyPlaybackState(): Promise<void> {
+    try {
+      const state = await this.spotifyApiWrapper.getPlaybackState();
+      SpotifySpeakerAccessory.CURRENT_STATE = state?.body;
+    } catch {
+      this.logger.error('Error setting spotify playback state');
     }
   }
 
@@ -134,11 +171,11 @@ export class HomebridgeSpotifySpeakerPlatform implements DynamicPlatformPlugin {
 
       const url = new URL(playlistUrl);
       const playlistId = url.pathname.split('/')[2];
-      this.log.debug(`Found playlistId: ${playlistId}`);
+      this.logger.debug(`Found playlistId: ${playlistId}`);
 
       return playlistId;
     } catch (error) {
-      this.log.error(
+      this.logger.error(
         `Failed to extract playlist ID, the plugin might behave in an unexpected way.
         Please check the configuration and provide a valid playlist URL`,
       );
@@ -149,7 +186,7 @@ export class HomebridgeSpotifySpeakerPlatform implements DynamicPlatformPlugin {
 
   private getDeviceConstructor(deviceType): typeof SpotifySpeakerAccessory | null {
     if (!deviceType) {
-      this.log.error('It is missing the `deviceType` in the configuration.');
+      this.logger.error('It is missing the `deviceType` in the configuration.');
       return null;
     }
 
@@ -157,14 +194,14 @@ export class HomebridgeSpotifySpeakerPlatform implements DynamicPlatformPlugin {
   }
 
   private async logAvailableSpotifyDevices(): Promise<void> {
-    const spotifyDevices = await this.spotifyApiWrapper.getMyDevices();
+    await this.setSpotifyDevices();
 
-    if (!spotifyDevices || spotifyDevices.length === 0) {
-      this.log.warn(
+    if (!SpotifySpeakerAccessory.DEVICES?.length) {
+      this.logger.warn(
         'No available spotify devices found, make sure that the speaker you configured is On and visible by Spotify Connect',
       );
     } else {
-      this.log.info('Available Spotify devices', spotifyDevices);
+      this.logger.info('Available Spotify devices', SpotifySpeakerAccessory.DEVICES);
     }
   }