+++ src/spotify-api-wrapper.test.ts
@@ -1,32 +1,16 @@
-import { API, Logger, PlatformConfig } from 'homebridge';
+import { API, PlatformConfig } from 'homebridge';
 import SpotifyWebApi from 'spotify-web-api-node';
+import { SpotifyDeviceNotFoundError } from './errors';
+import type { PluginLogger } from './plugin-logger';
 import { SpotifyApiWrapper } from './spotify-api-wrapper';
 import type { WebapiErrorBody } from './types';
 import { WebapiError } from './types';
-import { SpotifyDeviceNotFoundError } from './errors';
 
 jest.mock('spotify-web-api-node');
 const spotifyWebApiMocks = SpotifyWebApi as jest.MockedClass<typeof SpotifyWebApi>;
 
 beforeEach(() => {
   spotifyWebApiMocks.mockClear();
-});
-
-test('should re-attempt to retrieve device list once', async () => {
-  // given
-  const wrapper = getSpotifyApiWrapper();
-
-  const mockSpotifyWebApi = getMockedSpotifyWebApi();
-  mockSpotifyWebApi.getMyDevices
-    .mockRejectedValueOnce(new Error('something went wrong'))
-    .mockResolvedValueOnce(fakeGetMyDevicesResponse);
-
-  // when
-  const devices = await wrapper.getMyDevices();
-
-  // then
-  expect(devices?.length).toBe(1);
-  expect(mockSpotifyWebApi.getMyDevices).toHaveBeenCalledTimes(2);
 });
 
 test('should re-attempt wrapped request once on HTTP 404', async () => {
@@ -60,6 +44,7 @@ test('should throw SpotifyDeviceNotFoundError on second HTTP 404', async () => {
 
   // then
   await expect(playbackState).rejects.toThrow(SpotifyDeviceNotFoundError);
+  expect(mockSpotifyWebApi.getMyCurrentPlaybackState).toHaveBeenCalledTimes(2);
 });
 
 test('should refresh token on HTTP 401', async () => {
@@ -80,40 +65,41 @@ test('should refresh token on HTTP 401', async () => {
   expect(mockSpotifyWebApi.getMyCurrentPlaybackState).toHaveBeenCalledTimes(2);
 });
 
+test('should retry once for auth and once for device not found', async () => {
+  // given
+  const wrapper = getSpotifyApiWrapper();
+
+  const mockSpotifyWebApi = getMockedSpotifyWebApi();
+  mockSpotifyWebApi.getMyCurrentPlaybackState
+    .mockRejectedValueOnce(new WebapiError('test', {} as WebapiErrorBody, {}, 401, 'unauthorized'))
+    .mockRejectedValueOnce(new WebapiError('test', {} as WebapiErrorBody, {}, 404, 'device not found'))
+    .mockResolvedValueOnce(fakeCurrentPlaybackResponse);
+  mockSpotifyWebApi.refreshAccessToken.mockResolvedValueOnce(fakeAccessTokenResponse);
+
+  // when
+  const playbackState = await wrapper.getPlaybackState();
+
+  // then
+  expect(playbackState?.body.is_playing).toBe(false);
+  expect(mockSpotifyWebApi.getMyCurrentPlaybackState).toHaveBeenCalledTimes(3);
+});
+
 function getSpotifyApiWrapper(): SpotifyApiWrapper {
   return new SpotifyApiWrapper(
-    console as Logger,
+    console as unknown as PluginLogger,
     {
       spotifyAuthCode: '',
       spotifyClientId: '',
       spotifyClientSecret: '',
+      deviceNotFoundRetry: { enable: true, retryDelay: 1 },
     } as unknown as PlatformConfig,
     { user: { persistPath: () => '.' } } as API,
-    1,
   );
 }
 
 function getMockedSpotifyWebApi(): jest.Mocked<SpotifyWebApi> {
   return spotifyWebApiMocks.mock.instances[0] as jest.Mocked<SpotifyWebApi>;
 }
-
-const fakeGetMyDevicesResponse = {
-  body: {
-    devices: [
-      {
-        id: 'id',
-        is_active: false,
-        is_restricted: false,
-        is_private_session: false,
-        name: 'test',
-        type: 'type',
-        volume_percent: 0,
-      },
-    ],
-  },
-  headers: {},
-  statusCode: 200,
-};
 
 const fakeCurrentPlaybackResponse = {
   body: { is_playing: false } as SpotifyApi.CurrentPlaybackResponse,