+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.3.2] - 2023-08-06
+## [1.3.2] - 2023-08-08
+
+### Added
+
+- New plugin setting for debug logging. This is helpful for troubleshooting issues and when requesting support.
 
 ### Fixed
 
-- Automatically retry once on error because some Wi-Fi enabled Spotify speakers cannot be found on the first attempt after a period of inactivity.
+- New plugin settings to enable and configure one automatic retry because some Wi-Fi enabled Spotify speakers cannot be found when attempting to start a playlist after a period of inactivity. (#47)
 
 ## [1.3.1] - 2023-04-29